+++ core/CMakeLists.txt
@@ -11,7 +11,6 @@ add_subdirectory(timer)
 add_subdirectory(systemContent)
 add_subdirectory(fileManager)
 add_subdirectory(imports)
-add_subdirectory(semaphore)
 add_subdirectory(memory)
 add_subdirectory(dmem)
 
@@ -24,7 +23,6 @@ add_library(core SHARED
   $<TARGET_OBJECTS:videoout>
   $<TARGET_OBJECTS:fileManager>
   $<TARGET_OBJECTS:imports>
-  $<TARGET_OBJECTS:semaphore>
   $<TARGET_OBJECTS:memory>
   $<TARGET_OBJECTS:dmem>
 )
@@ -45,4 +43,4 @@ ADD_CUSTOM_COMMAND(TARGET core POST_BUILD
   ${CMAKE_CURRENT_BINARY_DIR}/core.lib
   ${CMAKE_BINARY_DIR}/lib)
 
-  install(FILES $<TARGET_PDB_FILE:core> DESTINATION debug OPTIONAL)
+install(FILES $<TARGET_PDB_FILE:core> DESTINATION debug OPTIONAL)
