+++ core/kernel/CMakeLists.txt
@@ -4,6 +4,7 @@ add_library(kernel OBJECT
   errors.cpp
   filesystem.cpp
   pthread.cpp
+  semaphore.cpp
 )
 
 add_dependencies(kernel third_party)
