--- core/semaphore/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_library(semaphore OBJECT
-  semaphore.cpp
-)
-
-add_dependencies(semaphore third_party)
