--- docs/uml/modules/simpleOverview.puml
@@ -1,17 +1,0 @@
-@startuml
-skinparam classAttributeIconSize 0
-
-package "Modules" {}
-
-class "psOff Emulator"{
-  +accessRuntimeLinker()
-  +accessInitParams()
-  +accessTimer()
-  +accessVideoOut()
-  +accessFileManager()
-  +accessGpuMemory()
-  +accessSystemContent()
-}
-
-Modules -> "psOff Emulator"
-@enduml
